Moved taggable model plugin to a separate 'base_taggable.rb' file.
