+++ vnet/lib/vnet/constants/translation.rb
@@ -4,5 +4,6 @@ module Vnet::Constants::Translation
   MODE_STATIC_ADDRESS = 'static_address'
   MODE_VNET_EDGE      = 'vnet_edge'
 
-  MODES = [ MODE_STATIC_ADDRESS, MODE_VNET_EDGE ]
+  MODES = [MODE_STATIC_ADDRESS,
+           MODE_VNET_EDGE]
 end
